FIX YOUR DAMN CODE EDX

+++ server/src/seeds/cleanDb.ts
@@ -3,11 +3,19 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    if (!db) {
+      throw new Error('fuck')
+    }
+    if (!models[modelName] || !models[modelName].db) {
+      throw new Error(`The model is possibly undefined`);
+    }
+
+    // ? is optional chaining, if it IS underfined this `?` could cause issues
+    const modelExists = await models[modelName]?.db?.db?.listCollections({
       name: collectionName
     }).toArray()
 
-    if (modelExists.length) {
+    if (modelExists && modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
